Prep tests, organize mocks, change page validation strategy

+++ lib/us_news_rankings/jobs/page_parser.rb
@@ -8,13 +8,16 @@ module UsNewsRankings
     end
 
     def perform
-      FileUtils.rm_rf(table_filepath)
-
-      File.open(table_filepath, 'w') do |file|
-        file.write(table.to_xhtml(indent: 2))
+      if page.valid?
+        FileUtils.rm_rf(table_filepath)
+        File.open(table_filepath, 'w') do |file|
+          file.write(table.to_xhtml(indent: 2))
+        end
+        FileUtils.rm_rf(page.filepath)
+        return true
+      else
+        return false
       end
-
-      FileUtils.rm_rf(page.filepath)
     end
 
     def table_filepath
